修改了test_config_manager_*.py文件的结构，尝试抽取出重复部分(未验证)
